Removing CUF runtime dependency with llvm::EnableABIBreakingChecks
